Reorganized Constants class
